+++ client/src/App.js
@@ -24,7 +24,7 @@ function App() {
 
   const handleLogout = () => {
     axios
-      .post("http://localhost:80/common/kakaosignout", {
+      .post("http://localhost:4000/common/kakaosignout", {
         userid: localStorage.getItem("userid"),
       })
       .then((res) => {
@@ -35,6 +35,25 @@ function App() {
         navigate("/");
       });
   };
+
+  const getGoogleToken = async (authorizationCode) => {
+    await axios({
+      url: "http://localhost:4000/oauth/google",
+      method: "post",
+      data: { authorizationCode },
+      withCredentials: true,
+    }).then((res) => {});
+  };
+
+  useEffect(() => {
+    const url = new URL(window.location.href);
+    console.log(`url ${url}`);
+    const authorizationCode = url.searchParams.get("code");
+    console.log(`authorizationCode ${authorizationCode}`);
+    if (authorizationCode) {
+      getGoogleToken(authorizationCode);
+    }
+  }, []);
 
   return (
     <>
